update: pro/createCard 리팩토링
